// src/types/ClientPersonalInfo.ts
export interface formData {
    participantName: string;
    dateOfBirth: string; // ISO format date string
    dateOfInitialAssessment: string; // ISO format date string
    dateOfCurrentReassessment: string; // ISO format date string
    currentFamilyStructure: string;
    otherHouseholdDetails: string;
    parentName: string;
    parentPhone: string;
    parentEmail: string;
  }